refactor trajectory
