+++ src/mrseq/utils/__init__.py
@@ -1,5 +1,7 @@
 from mrseq.utils.sequence_helper import find_gx_flat_time_on_adc_raster, round_to_raster, write_sequence
 from mrseq.utils.system_defaults import sys_defaults
 from mrseq.utils.vds import variable_density_spiral_trajectory
-from mrseq.utils.trajectory import spiral_acquisition
+from mrseq.utils.spiral_sampling import spiral_acquisition
+from mrseq.utils.cartesian_sampling import cartesian_phase_encoding
+from mrseq.utils.MultiEchoAcquisition import MultiEchoAcquisition
 from mrseq.utils.ismrmrd import create_header, combine_ismrmrd_files
